fix wrong data type in tntdb::Decimal::getUnsignedShort()
git-svn-id: svn://svn.code.sf.net/p/tntdb/code/trunk@255 8aebba32-5d12-0410-a889-bbfed4e15866

+++ tntdb/include/tntdb/decimal.h
@@ -202,7 +202,7 @@ namespace tntdb
     /// @return unsigned result if the result will fit.
     /// @throw std::overflow_error if the result will not fit
     unsigned short getUnsignedShort() const throw(std::overflow_error)
-      { return getInteger<unsigned>(); }
+      { return getInteger<unsigned short>(); }
 
     /// Return this decimal number rounded as a C++ unsigned.
     /// @return unsigned result if the result will fit.
